refactor: made MillionsChart more modular
