+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/view/components/MillionsChart/MillionsChart.java
@@ -24,25 +24,14 @@ public class MillionsChart extends MillionsView {
   private NumberAxis xAxis;
   private ChartMode chartMode = ChartMode.MAX;
 
-  private String name;
-  private BigDecimal value;
-  private BigDecimal change;
-
-  private Label nameLabel;
-  private Label valueLabel;
-  private Label changeLabel;
-
   private List<Button> filterButtons;
 
-  public MillionsChart(String name, BigDecimal value, BigDecimal change) {
+  public MillionsChart() {
     getStylesheets().add(Objects.requireNonNull(
       getClass().getResource("/styles/chart.css")
     ).toExternalForm());
 
     this.series = new XYChart.Series<>();
-    this.name = name;
-    this.value = value;
-    this.change = change;
 
     getChildren().add(build());
   }
@@ -56,67 +45,11 @@ public class MillionsChart extends MillionsView {
     VBox chartContainer = new VBox();
     chartContainer.getStyleClass().add("chart-container");
     chartContainer.getChildren().addAll(
-      createInfoBar(),
       filters(),
       createChart()
     );
 
     return chartContainer;
-  }
-
-  private BorderPane createInfoBar() {
-    nameLabel = new Label(name);
-    valueLabel = new Label(value.toString());
-    changeLabel = new Label(change.toString() + "%");
-
-    nameLabel.getStyleClass().add("name");
-    valueLabel.getStyleClass().add("price");
-    changeLabel.getStyleClass().add("change");
-
-    updateChange();
-
-    HBox valueSection = new HBox();
-    valueSection.getStyleClass().add("value-section");
-    valueSection.getChildren().addAll(
-      valueLabel,
-      changeLabel
-    );
-
-    BorderPane infoBar = new BorderPane();
-    infoBar.getStyleClass().add("info-bar");
-    infoBar.setLeft(nameLabel);
-    infoBar.setRight(valueSection);
-
-    return infoBar;
-  }
-
-  public void updateInfoBar(String name, BigDecimal value, BigDecimal change) {
-    this.name = name;
-    this.value = value;
-    this.change = change;
-
-    nameLabel.setText(name);
-    valueLabel.setText(value.toString());
-    changeLabel.setText(change + "%");
-
-    updateChange();
-  }
-
-  private void updateChange() {
-    changeLabel.getStyleClass().removeAll("positive-change", "negative-change", "zero-change");
-
-    if (change.signum() > 0) {
-      changeLabel.setText("+" + change + "%");
-      changeLabel.getStyleClass().add("positive-change");
-    }
-    else if (change.signum() < 0) {
-      changeLabel.setText(change + "%");
-      changeLabel.getStyleClass().add("negative-change");
-    }
-    else {
-      changeLabel.setText(change + "%");
-      changeLabel.getStyleClass().add("zero-change");
-    }
   }
 
   private AreaChart<Number, Number> createChart() {
