+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/view/tradingpage/TradingPageView.java
@@ -4,14 +4,17 @@ import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.Pane;
 import org.edu.ntnu.idatt2003.group49.millions.controller.Navigator;
 import org.edu.ntnu.idatt2003.group49.millions.view.MillionsView;
+import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsChart.MillionsChart;
 import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsTable.TradingTable;
 import org.edu.ntnu.idatt2003.group49.millions.view.tradingpage.components.StockInfo;
 
 public class TradingPageView extends MillionsView {
   private final TradingTable tradingTable;
+  private final MillionsChart chart;
 
-  public TradingPageView(TradingTable tradingTable) {
+  public TradingPageView(TradingTable tradingTable, MillionsChart chart) {
     this.tradingTable = tradingTable;
+    this.chart = chart;
 
     getChildren().add(build());
   }
@@ -20,7 +23,7 @@ public class TradingPageView extends MillionsView {
   protected Pane build() {
     BorderPane body = new BorderPane();
     body.setLeft(tradingTable);
-    body.setRight(new StockInfo());
+    body.setRight(new StockInfo(chart));
     return body;
   }
 }
