+++ src/main/java/org/edu/ntnu/idatt2003/group49/millions/view/tradingpage/components/StockInfo.java
@@ -4,16 +4,24 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.Pane;
 import javafx.scene.layout.VBox;
 import org.edu.ntnu.idatt2003.group49.millions.view.MillionsView;
+import org.edu.ntnu.idatt2003.group49.millions.view.components.MillionsChart.MillionsChart;
 
 public class StockInfo extends MillionsView {
-  public StockInfo() {
+  private final MillionsChart chart;
+
+  public StockInfo(MillionsChart chart) {
+    this.chart = chart;
+
     getChildren().add(build());
   }
 
   @Override
   protected Pane build() {
     VBox body = new VBox();
-    body.getChildren().add(new Label("Stock Info"));
+    body.getChildren().addAll(
+      new Label("Stock Info"),
+      chart
+    );
     return body;
   }
 }
